wip: evolution logic refactor
