+++ app/core/evolution-logic/condition-evolution-handler.ts
@@ -1,0 +1,34 @@
+import type Player from "../../models/colyseus-models/player"
+import type { Pokemon } from "../../models/colyseus-models/pokemon"
+import { Item } from "../../types"
+import type { ConditionEvolutionRule } from "../../types/EvolutionRules"
+import { EvolutionHandler } from "./evolution-handler"
+
+type EvolutionCondition = (
+  pokemon: Pokemon,
+  player: Player,
+  stageLevel: number
+) => boolean
+
+export class ConditionEvolutionHandler extends EvolutionHandler {
+  condition: EvolutionCondition
+  constructor(evolutionRule: ConditionEvolutionRule) {
+    super(evolutionRule.divergentEvolution)
+    this.condition = evolutionRule.condition
+  }
+
+  canEvolve(pokemon: Pokemon, player: Player, stageLevel: number): boolean {
+    if (pokemon.items.has(Item.EVIOLITE)) return false
+    if (player.board.has(pokemon.id) === false) return false
+    return this.condition(pokemon, player, stageLevel)
+  }
+
+  evolve(pokemon: Pokemon, player: Player, stageLevel: number): Pokemon {
+    const pokemonEvolutionName = this.getEvolution(pokemon, player, stageLevel)
+    const pokemonEvolved = player.transformPokemon(
+      pokemon,
+      pokemonEvolutionName
+    )
+    return pokemonEvolved
+  }
+}
